REF: DataFrame delitem, take, pop, filter tests (#33109)
