+++ pandas/tests/frame/test_axis_select_reindex.py
@@ -12,16 +12,6 @@ class TestDataFrameSelectReindex:
     # These are specific reindex-based tests; other indexing tests should go in
     # test_indexing
 
-    def test_delitem_col_still_multiindex(self):
-        arrays = [["a", "b", "c", "top"], ["", "", "", "OD"], ["", "", "", "wx"]]
-
-        tuples = sorted(zip(*arrays))
-        index = MultiIndex.from_tuples(tuples)
-
-        df = DataFrame(np.random.randn(3, 4), columns=index)
-        del df[("a", "", "")]
-        assert isinstance(df.columns, MultiIndex)
-
     def test_merge_join_different_levels(self):
         # GH 9455
 
@@ -409,220 +399,6 @@ class TestDataFrameSelectReindex:
         result = df1 - df1.mean()
         expected = df2 - df2.mean()
         tm.assert_frame_equal(result, expected)
-
-    def test_filter(self, float_frame, float_string_frame):
-        # Items
-        filtered = float_frame.filter(["A", "B", "E"])
-        assert len(filtered.columns) == 2
-        assert "E" not in filtered
-
-        filtered = float_frame.filter(["A", "B", "E"], axis="columns")
-        assert len(filtered.columns) == 2
-        assert "E" not in filtered
-
-        # Other axis
-        idx = float_frame.index[0:4]
-        filtered = float_frame.filter(idx, axis="index")
-        expected = float_frame.reindex(index=idx)
-        tm.assert_frame_equal(filtered, expected)
-
-        # like
-        fcopy = float_frame.copy()
-        fcopy["AA"] = 1
-
-        filtered = fcopy.filter(like="A")
-        assert len(filtered.columns) == 2
-        assert "AA" in filtered
-
-        # like with ints in column names
-        df = DataFrame(0.0, index=[0, 1, 2], columns=[0, 1, "_A", "_B"])
-        filtered = df.filter(like="_")
-        assert len(filtered.columns) == 2
-
-        # regex with ints in column names
-        # from PR #10384
-        df = DataFrame(0.0, index=[0, 1, 2], columns=["A1", 1, "B", 2, "C"])
-        expected = DataFrame(
-            0.0, index=[0, 1, 2], columns=pd.Index([1, 2], dtype=object)
-        )
-        filtered = df.filter(regex="^[0-9]+$")
-        tm.assert_frame_equal(filtered, expected)
-
-        expected = DataFrame(0.0, index=[0, 1, 2], columns=[0, "0", 1, "1"])
-        # shouldn't remove anything
-        filtered = expected.filter(regex="^[0-9]+$")
-        tm.assert_frame_equal(filtered, expected)
-
-        # pass in None
-        with pytest.raises(TypeError, match="Must pass"):
-            float_frame.filter()
-        with pytest.raises(TypeError, match="Must pass"):
-            float_frame.filter(items=None)
-        with pytest.raises(TypeError, match="Must pass"):
-            float_frame.filter(axis=1)
-
-        # test mutually exclusive arguments
-        with pytest.raises(TypeError, match="mutually exclusive"):
-            float_frame.filter(items=["one", "three"], regex="e$", like="bbi")
-        with pytest.raises(TypeError, match="mutually exclusive"):
-            float_frame.filter(items=["one", "three"], regex="e$", axis=1)
-        with pytest.raises(TypeError, match="mutually exclusive"):
-            float_frame.filter(items=["one", "three"], regex="e$")
-        with pytest.raises(TypeError, match="mutually exclusive"):
-            float_frame.filter(items=["one", "three"], like="bbi", axis=0)
-        with pytest.raises(TypeError, match="mutually exclusive"):
-            float_frame.filter(items=["one", "three"], like="bbi")
-
-        # objects
-        filtered = float_string_frame.filter(like="foo")
-        assert "foo" in filtered
-
-        # unicode columns, won't ascii-encode
-        df = float_frame.rename(columns={"B": "\u2202"})
-        filtered = df.filter(like="C")
-        assert "C" in filtered
-
-    def test_filter_regex_search(self, float_frame):
-        fcopy = float_frame.copy()
-        fcopy["AA"] = 1
-
-        # regex
-        filtered = fcopy.filter(regex="[A]+")
-        assert len(filtered.columns) == 2
-        assert "AA" in filtered
-
-        # doesn't have to be at beginning
-        df = DataFrame(
-            {"aBBa": [1, 2], "BBaBB": [1, 2], "aCCa": [1, 2], "aCCaBB": [1, 2]}
-        )
-
-        result = df.filter(regex="BB")
-        exp = df[[x for x in df.columns if "BB" in x]]
-        tm.assert_frame_equal(result, exp)
-
-    @pytest.mark.parametrize(
-        "name,expected",
-        [
-            ("a", DataFrame({"a": [1, 2]})),
-            ("a", DataFrame({"a": [1, 2]})),
-            ("あ", DataFrame({"あ": [3, 4]})),
-        ],
-    )
-    def test_filter_unicode(self, name, expected):
-        # GH13101
-        df = DataFrame({"a": [1, 2], "あ": [3, 4]})
-
-        tm.assert_frame_equal(df.filter(like=name), expected)
-        tm.assert_frame_equal(df.filter(regex=name), expected)
-
-    @pytest.mark.parametrize("name", ["a", "a"])
-    def test_filter_bytestring(self, name):
-        # GH13101
-        df = DataFrame({b"a": [1, 2], b"b": [3, 4]})
-        expected = DataFrame({b"a": [1, 2]})
-
-        tm.assert_frame_equal(df.filter(like=name), expected)
-        tm.assert_frame_equal(df.filter(regex=name), expected)
-
-    def test_filter_corner(self):
-        empty = DataFrame()
-
-        result = empty.filter([])
-        tm.assert_frame_equal(result, empty)
-
-        result = empty.filter(like="foo")
-        tm.assert_frame_equal(result, empty)
-
-    def test_filter_regex_non_string(self):
-        # GH#5798 trying to filter on non-string columns should drop,
-        #  not raise
-        df = pd.DataFrame(np.random.random((3, 2)), columns=["STRING", 123])
-        result = df.filter(regex="STRING")
-        expected = df[["STRING"]]
-        tm.assert_frame_equal(result, expected)
-
-    def test_take(self, float_frame):
-        # homogeneous
-        order = [3, 1, 2, 0]
-        for df in [float_frame]:
-
-            result = df.take(order, axis=0)
-            expected = df.reindex(df.index.take(order))
-            tm.assert_frame_equal(result, expected)
-
-            # axis = 1
-            result = df.take(order, axis=1)
-            expected = df.loc[:, ["D", "B", "C", "A"]]
-            tm.assert_frame_equal(result, expected, check_names=False)
-
-        # negative indices
-        order = [2, 1, -1]
-        for df in [float_frame]:
-
-            result = df.take(order, axis=0)
-            expected = df.reindex(df.index.take(order))
-            tm.assert_frame_equal(result, expected)
-
-            result = df.take(order, axis=0)
-            tm.assert_frame_equal(result, expected)
-
-            # axis = 1
-            result = df.take(order, axis=1)
-            expected = df.loc[:, ["C", "B", "D"]]
-            tm.assert_frame_equal(result, expected, check_names=False)
-
-        # illegal indices
-        msg = "indices are out-of-bounds"
-        with pytest.raises(IndexError, match=msg):
-            df.take([3, 1, 2, 30], axis=0)
-        with pytest.raises(IndexError, match=msg):
-            df.take([3, 1, 2, -31], axis=0)
-        with pytest.raises(IndexError, match=msg):
-            df.take([3, 1, 2, 5], axis=1)
-        with pytest.raises(IndexError, match=msg):
-            df.take([3, 1, 2, -5], axis=1)
-
-    def test_take_mixed_type(self, float_string_frame):
-
-        # mixed-dtype
-        order = [4, 1, 2, 0, 3]
-        for df in [float_string_frame]:
-
-            result = df.take(order, axis=0)
-            expected = df.reindex(df.index.take(order))
-            tm.assert_frame_equal(result, expected)
-
-            # axis = 1
-            result = df.take(order, axis=1)
-            expected = df.loc[:, ["foo", "B", "C", "A", "D"]]
-            tm.assert_frame_equal(result, expected)
-
-        # negative indices
-        order = [4, 1, -2]
-        for df in [float_string_frame]:
-
-            result = df.take(order, axis=0)
-            expected = df.reindex(df.index.take(order))
-            tm.assert_frame_equal(result, expected)
-
-            # axis = 1
-            result = df.take(order, axis=1)
-            expected = df.loc[:, ["foo", "B", "D"]]
-            tm.assert_frame_equal(result, expected)
-
-    def test_take_mixed_numeric(self, mixed_float_frame, mixed_int_frame):
-        # by dtype
-        order = [1, 2, 0, 3]
-        for df in [mixed_float_frame, mixed_int_frame]:
-
-            result = df.take(order, axis=0)
-            expected = df.reindex(df.index.take(order))
-            tm.assert_frame_equal(result, expected)
-
-            # axis = 1
-            result = df.take(order, axis=1)
-            expected = df.loc[:, ["B", "C", "A", "D"]]
-            tm.assert_frame_equal(result, expected)
 
     def test_reindex_boolean(self):
         frame = DataFrame(
