+++ pandas/tests/frame/test_mutate_columns.py
@@ -1,9 +1,7 @@
-import re
-
 import numpy as np
 import pytest
 
-from pandas import DataFrame, MultiIndex, Series
+from pandas import DataFrame, Series
 import pandas._testing as tm
 
 # Column add, remove, delete.
@@ -83,69 +81,3 @@ class TestDataFrameMutateColumns:
         df["X"] = ["x", "y", "z"]
         exp = DataFrame(data={"X": ["x", "y", "z"]}, index=["A", "B", "C"])
         tm.assert_frame_equal(df, exp)
-
-    def test_delitem(self, float_frame):
-        del float_frame["A"]
-        assert "A" not in float_frame
-
-    def test_delitem_multiindex(self):
-        midx = MultiIndex.from_product([["A", "B"], [1, 2]])
-        df = DataFrame(np.random.randn(4, 4), columns=midx)
-        assert len(df.columns) == 4
-        assert ("A",) in df.columns
-        assert "A" in df.columns
-
-        result = df["A"]
-        assert isinstance(result, DataFrame)
-        del df["A"]
-
-        assert len(df.columns) == 2
-
-        # A still in the levels, BUT get a KeyError if trying
-        # to delete
-        assert ("A",) not in df.columns
-        with pytest.raises(KeyError, match=re.escape("('A',)")):
-            del df[("A",)]
-
-        # behavior of dropped/deleted MultiIndex levels changed from
-        # GH 2770 to GH 19027: MultiIndex no longer '.__contains__'
-        # levels which are dropped/deleted
-        assert "A" not in df.columns
-        with pytest.raises(KeyError, match=re.escape("('A',)")):
-            del df["A"]
-
-    def test_pop(self, float_frame):
-        float_frame.columns.name = "baz"
-
-        float_frame.pop("A")
-        assert "A" not in float_frame
-
-        float_frame["foo"] = "bar"
-        float_frame.pop("foo")
-        assert "foo" not in float_frame
-        assert float_frame.columns.name == "baz"
-
-        # gh-10912: inplace ops cause caching issue
-        a = DataFrame([[1, 2, 3], [4, 5, 6]], columns=["A", "B", "C"], index=["X", "Y"])
-        b = a.pop("B")
-        b += 1
-
-        # original frame
-        expected = DataFrame([[1, 3], [4, 6]], columns=["A", "C"], index=["X", "Y"])
-        tm.assert_frame_equal(a, expected)
-
-        # result
-        expected = Series([2, 5], index=["X", "Y"], name="B") + 1
-        tm.assert_series_equal(b, expected)
-
-    def test_pop_non_unique_cols(self):
-        df = DataFrame({0: [0, 1], 1: [0, 1], 2: [4, 5]})
-        df.columns = ["a", "b", "a"]
-
-        res = df.pop("a")
-        assert type(res) == DataFrame
-        assert len(res) == 2
-        assert len(df.columns) == 1
-        assert "b" in df.columns
-        assert "a" not in df.columns
-        assert len(df.index) == 2
